优化删除远端服务器上的好友关系

+++ src/cn/ucai/superwechat/activity/MainActivity.java
@@ -582,32 +582,28 @@ public class MainActivity extends BaseActivity implements EMEventListener {
 		public void onContactDeleted(final List<String> usernameList) {
             Log.e(TAG,"onContactDeleted,usernameList="+usernameList);
 			// 被删除
-			Map<String, User> localUsers = ((DemoHXSDKHelper)HXSDKHelper.getInstance()).getContactList();
-            String currentUserName = SuperWeChatApplication.getInstance().getUserName();
-            List<String> toDelUserName = new ArrayList<String>();
-            for (String username : usernameList) {
-                Log.e(TAG,"onContactDeleted,username="+username);
-				localUsers.remove(username);
-                toDelUserName.add(username);
 
-				userDao.deleteContact(username);
-				inviteMessgeDao.deleteMessage(username);
-			}
-            for(final String name:toDelUserName){
+            String currentUserName = SuperWeChatApplication.getInstance().getUserName();
+            for (final String username : usernameList) {
+                Log.e(TAG,"onContactDeleted,username="+username);
                 final OkHttpUtils2<Result> utils = new OkHttpUtils2<Result>();
                 utils.setRequestUrl(I.REQUEST_DELETE_CONTACT)
                         .addParam(I.Contact.USER_NAME,currentUserName)
-                        .addParam(I.Contact.CU_NAME,name)
+                        .addParam(I.Contact.CU_NAME,username)
                         .targetClass(Result.class)
                         .execute(new OkHttpUtils2.OnCompleteListener<Result>() {
                             @Override
                             public void onSuccess(Result result) {
-                                Map<String, UserAvatar> userMap = SuperWeChatApplication.getInstance().getUserMap();
-                                List<UserAvatar> userList = SuperWeChatApplication.getInstance().getUserList();
-                                UserAvatar u = userMap.get(name);
-                                userList.remove(u);
-                                userMap.remove(name);
-                                sendStickyBroadcast(new Intent("update_contact_list"));
+                                if(result.isRetMsg()) {
+                                    ((DemoHXSDKHelper) HXSDKHelper.getInstance()).getContactList().remove(username);
+                                    UserAvatar u = SuperWeChatApplication.getInstance().getUserMap().get(username);
+                                    SuperWeChatApplication.getInstance().getUserList().remove(u);
+                                    SuperWeChatApplication.getInstance().getUserMap().remove(username);
+
+                                    userDao.deleteContact(username);
+                                    inviteMessgeDao.deleteMessage(username);
+                                    sendStickyBroadcast(new Intent("update_contact_list"));
+                                }
                             }
 
                             @Override
@@ -615,7 +611,7 @@ public class MainActivity extends BaseActivity implements EMEventListener {
                                 Log.e(TAG,"error="+error);
                             }
                         });
-            }
+			}
 			runOnUiThread(new Runnable() {
 				public void run() {
 					// 如果正在与此用户的聊天页面
